all: Use system Libreswan NSS certificate and key database
nssdir= setting in the Libreswan config section of the generated ipsec.conf
file was being ignored. Unfortunately only option seems to be to use the
system Libreswan NSS database.

The additional code handles the location of the NSS database as well as
parsing the Libreswan NSS DB password file if it exists to authenticate
the NSS database if required.

It assumes the system Libreswan NSS database has previously been initialized
and will produce an error if it wasn't.