all: Remove Prevalent Algorithms button, overide default algorithms
Make the phase 1 & 2 proposals provided by the Prevalent Algorithms button
the default. The algorithms are a merge of Windows 10 and macOS/iOS/iPadOS
L2TP clients' IKEv1 proposals.

It was a compromise for backwards compatibilty with legacy VPN servers which
would work with Win10 and macOS/iOS/iPadOS L2TP/IPsec clients, but didn't
with  libreswan or strongSwan default set of proposals.