all: Remove Prevalent Algorithms button, overide default algorithms
Make the phase 1 & 2 proposals that were provided by the Prevalent
Algorithms button the default.

The algorithms are a merge of Windows 10 and macOS/iOS/iPadOS
L2TP clients' IKEv1 proposals. The weakest proposals that weren't common
to both Win10 and iOS were dropped, but all of the strongest ones were
kept.

It was a compromise for backwards compatibilty with legacy VPN servers which
would work with Win10 and macOS/iOS/iPadOS L2TP/IPsec clients, but didn't
with  libreswan or strongSwan default set of proposals.