README.md: remove info about potential GPL/OpenSSL license conflict
The GPL/OpenSSL license conflict seems to be an issue that affects Debian,
Ubuntu and derivates. Most Linux distros ship a potsgresql package that links
OpenSSL with GNU readline library (GPLv3), so they seem to implictly consider
OpenSSL to be a  "System Library" and so exempt from the conflict per the
"System Library Exception"

The GPL/OpenSSL license conflict is described more thoroughly on the
releases download page.