build: Use a minimum gettext requirement rather than an exact one
In configure.ac, specify the minimum gettext version we require, rather
than the exact one. This fixes a situation where the autoconf macros
used for gettext will be the latest available on the system (for
example, 0.20); but the copied-in Makefile.in.in will be for the exact
version specified in configure.ac (in this case, 0.19).

In that situation, the gettext build rules will error out at `make` time
with the message:
   *** error: gettext infrastructure mismatch: using a Makefile.in.in
   from gettext version 0.19 but the autoconf macros are from gettext
   version 0.20

This patch is based on
https://gitlab.gnome.org/GNOME/geocode-glib/commit/6ba0e38fd1987be0de8432afd93b9a5c1f487b63