properties: modernize the layout
This simplifies the layout with use use of GtkGrids in place of of nested
GtkBox and GtkAlignment widgets. The latter widget class is deprecated
and gone from Gtk4.

While at that, it modernizes the alignment of the labels to match what
is used withing GNOME Control Center and newer versions of
nm-connection-editor.