service: check strongswan CHILD_SA state is INSTALLED
Checking the output of `/usr/sbin/ipsec status` for "ESTABLISHED" string is
not enough to determine if the strongswan connection was sucessful, as
"ESTABLISHED" could be for a IKE_SA, not a CHILD_SA, which is required to
transport traffic.

Now also checks for "INSTALLED, TRANSPORT" in the ipsec status output.