Add ipsec-left vpn-data option
By default a value of %defaultroute is used for the "left" parameter of
the IPSec daemon configuration. However that is not suitable on a
machine that needs to establish the IPSec connection via an interface that
does not correspond to the default route.

Add a "ipsec-left" option for vpn-data that allows a custom value to be
passed as the ipsec.conf "left" option for these more advanced setup
scenarios.