service: determine correct local IP address for IPSec
The ipsec %defaultroute keyword is used to have Strongswan/Libreswan
use the local address of the default route network interface at the point
of startup or configuration reload.  It's probably a fair bet that this
would be the interface used to connect to the VPN gateway, but that's
not guaranteed. To support the latter, determine the IP address
directly by connecting to the gateway and using the getsockname(2)
system call to derive the local socket address.

Adapted from:
https://github.com/katalix/NetworkManager-l2tp/commit/2d1720e17ddf9b7a7d02399e6a1eacb747080cd8
Credits to https://github.com/katalix and https://github.com/tomparkin
for the original work.