service: immediately leave validate_one_property() upon validation error
When setting an invalid property value (e.g. require-mppe: foobar), GLib
prints the following error:

    GError set over the top of a previous GError or uninitialized memory.
    This indicates a bug in someone's code. You must ensure an error is NULL
    before it's set..

This happens because the break statement that was apparently meant to short
circuit the outer for loop just exits from the inner case statement.
Eventually we fall through its end and attempt to overwrite the previously set
error as if the property wasn't found.

Based on following NetworkManager-pptp merge request:
https://gitlab.gnome.org/GNOME/NetworkManager-pptp/-/merge_requests/18