shared/vpn-plugin-utils: split editor plugin lookup and load
nm-connection-editor (and presumably the Control Center) expects the
nm_vpn_editor_plugin_factory() to fail if the editor plugin (the thing
that goes into the *-gnome subpackage in Fedora) is not installed.

However, factory() never fails, because the plugin is checked for
existence only when get_editor() is called.

Based on following NetworkManager-libreswan commit:
https://gitlab.gnome.org/GNOME/NetworkManager-libreswan/-/commit/683c5c1689a472ef5dc582dd1168cc0957ca63cb

Which comes from NetworkManager.git commit
'contrib/nm-vpn-plugin-utils: split editor plugin lookup and load'.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/2037