editor-plugin: check for existence of the editor
Currently, the users that use the GUI editor have no way of checking if
it is installed short of actually attempting to load it (which would be
slow and cumbresome).

The plugin factory always succeeds, and rightly so, because it could
still be providing import()/export() that was added with
NetworkManager-1.40 or so.

Let's check the presence of editor (the GUI part) upon plugin load and
indicate the results via the capabilities. Needs NetworkManager-1.52 to
work properly.

Based on following NetworkManager-libreswan commit:
https://gitlab.gnome.org/GNOME/NetworkManager-libreswan/-/commit/19f659eb6b24c98973eae8792f63c97122c2372b