auth-dialog: set locale before parsing options
If we don't do this, then the locale defaults to "C" which causes
subsequent call to g_option_context_parse() reject wide characters
which might be present in password prompt.

This shifts the burden of ensuring the string is encoded consistently
with locale set in environment to our caller (nm-applet or gnome-shell).
That is probably fine, and nowadays it's all UTF-8 anyway.

Note that when the auth dialog is linked with gtk3 this appears to be
done by Gtk's ctors -- that is no longer the case with gtk4.

https://bugzilla.redhat.com/show_bug.cgi?id=2388442

based on following NetworkManager-pptp commit:

https://gitlab.gnome.org/GNOME/NetworkManager-pptp/-/merge_requests/21/diffs?commit_id=d433a883ae961a350652dff4a4b05f37e345755e